Support build artifacts for external data files.
This change adds support of declaring build artifacts as external
data files in the following JSON format:

{
  "type": "artifact",
  "name": "license_credits.html"
}

If this JSON is found as *.external in data/ directory, the file at

  gs://chromeos-image-archive/$BUILDER_PATH/license_credits.html

is downloaded to the DUT, where $BUILDER_PATH is a string found as
CHROMEOS_RELEASE_BUILDER_PATH in /etc/lsb-release. Afterwards,
those files can be accessible in tests with s.DataPath just like
other static external data files.

BUG=chromium:943383
TEST=tast run -buildlocalrunner DUT example.DataFiles
TEST=fast_build.sh -T

Change-Id: I70dc14c9a2a7a70bd5986c7c3677b4fac4e0e73e