tast: tast-lint supports symlinks.
tast-lint: when parsing the CL, read the content of the symlink instead
of doing "git show". The problem with "git show" is that it returns the
path to the linked file, and not its content.

Also, files inside "/gen" and "/genutil" folders are not considered
tests. This allows using the forbidden imports like
"github.com/pkg/errors" in the generators.

BUG=None
TEST=./fast_build.sh
TEST=repo upload --cbr . (this cl: https://crrev.com/c/1533103)
     and CL passed the verification

Change-Id: Ic3fe1fb955c59b84883bcd887b4d1c101ca38537