tast: tast-lint supports symlinks to Go files.
Add support for symlinks to .go files in tast-lint.
Currently we are using "git show" to parse the diff in the CL.
And for symlinks, "git show" fails because it just returns the path
it links to, instead of its conent.
eg: "git show" of a newly added symlink will return:
   +../../../genutil/util.go

This CL treats symlinks differently and uses "git cat-file" to return
the content of symlinks.

Also, files inside "/gen" and "/genutil" folders are not considered
tests. This allows using the forbidden imports like
"github.com/pkg/errors" in the generators.

BUG=None
TEST=./fast_build.sh
TEST=repo upload --cbr . (this cl: https://crrev.com/c/1533103)
     and CL passed the verification

Change-Id: Ic3fe1fb955c59b84883bcd887b4d1c101ca38537