tast: Kill test bundles and related processes.
Update test runners to create a new session for each bundle
and kill all processes in the session (including processes
started by tests) after the bundle exits.

Also kill the session if the read end of stdout is closed
while tests are running, as this typically indicates that
the SSH connection was dropped. Otherwise, the processes can
hang around forever.

BUG=chromium:945626
TEST=added unit tests

Change-Id: I5e751ba828e21d0944a305d9a9eba64c16275095