tast: Introduce testing.Sleep.
It is forbidden by tast-lint to use time.Sleep in test codes for
two reasons: it does not honor context timeout, and also sleeping
can be a cause of flakiness. However we got some legitimate use
cases of sleep (like sending keyboard events) in more and more
tests, so it's good time to introduce a utility function for it.

BUG=chromium:951690
TEST=fast_build.sh -T

Change-Id: Ic78e6ca412c17632d06623fb23a87176ee7cd99f