tast: Check that bundle package isn't installed for host.
Make "tast run -build ..." check that the bundle package
(e.g. chromeos-base/tast-local-tests-cros-9999) isn't
installed in the host sysroot. Otherwise, the emerge command
that we run to check that dependencies are installed and
up-to-date ignores --onlydeps-with-rdeps=n and incorrectly
checks runtime dependencies (due to what looks like an
upstream bug: https://bugs.gentoo.org/682472).

Also make a small tweak to make "tast run" avoid reprinting
the error message if we didn't actually start running tests,
since the error message is probably already the last line of
output in that case.

BUG=chromium:940761
TEST=manual: checked that "tast run -build ..." prints an
     error after running "sudo emerge --nodeps
     tast-local-tests-cros"

Change-Id: I9923663fbbc5c986456af25b70b633ceff552002