tast: Add -stdin flag to "tast list".
To make it possible to decide which tests to run from
outside of Tast (e.g. in a scheduling system), add an -stdin
flag to the "list" subcommand that can be used to read JSON
arrays of testing.Test structs (e.g. as produced by bundles'
-dumptests flag) from stdin instead of fetching tests from a
DUT. Multiple arrays can be passed to make it easier to
provide tests from multiple bundles.

This change also adds a new NoFinalize option that can be
passed to testing.NewRegistry to indicate that added tests
should not be validated or auto-filled (since this change
requires adding serialized tests that were already finalized
in the bundle).

BUG=chromium:950347
TEST=added/updated unit tests; also manually verified that
     it works

Change-Id: Ide5d00959f5fc76e511b457b4cf7442e1fb2d07f