tast: Add -resume to "tast run".
Add a -resume flag that can be passed to "tast run" to
instruct it to reconnect to the DUT if needed and try to run
the remaining tests after a run error is encountered
mid-test (usually due to a bundle crash, DUT crash, or lost
SSH connection). The flag is off by default.

BUG=chromium:889119
TEST=updated unit tests

Change-Id: Ibf8d0359ca954658baa53eb6e66d4f94ce343a5c