tast: Report results for unfinished tests.
Make resultsHandler return TestResult entries for
started-but-not-finished tests. Previously, results were
only returned for completed tests, with incomplete tests not
included in results.json.

Also flush the in-progress test (if any) to
streamed_results.jsonl before closing the file to ensure
that all test errors will be recorded if we encounter a run
error mid-test.

BUG=chromium:889119
TEST=updated unit tests; also manually ran a test that made
     the test bundle crash and checked that it is reported
     as started but not finished in results.json

Cq-Depend: chromium:1595149
Change-Id: Ia6b62866e9021646e2ca4f99af2fe7b63b029985