tast: Add host.Forwarder.
Add a Forwarder class to the host package to facilitate
forwarding TCP connections between a local port and a remote
address by way of a remote SSH server.

This is based heavily on jeffcarp@'s ServodProxy change at
https://crrev.com/c/1542128. I'm adding a generic class to
the host package instead of making this Servo-specific since
I suspect that similar functionality will be needed
elsewhere in the future.

BUG=chromium:942366
TEST=added unit tests; also manually tested by modifying
     example.ServoEcho test to communicate with a servod
     instance in the lab by way of a forwarded connection

Change-Id: I2931852c7bbf88c1c5e18ebaef862b61ef3d35cb