Update patch set 2
Patch Set 2:

i'm curious to hear nya@'s thoughts about this, but in general, i'm leaning toward not adding a dedicated crosconfig package until there are additional use cases for it. it seems impossible to know what sort of interface is right at this point, especially with regard to defining property names in the package vs. having the caller pass them in.

is your initial intent to just use this for a single sensors test? if so, i'd recommend moving the code that runs cros_config and interprets its output into that test file for now. it's straightforward to introduce a new crosconfig package in the future once we have a better idea about the functionality that it needs to expose. if this crosconfig package is added now, it'll me harder to change its exported interface in the future -- at the very least, you'll need to update your test, which will require making simultaneous changes to the tast and tast-tests repositories.

Patch-set: 2