tast: Refactor code for starting ephemeral devservers.
Make the run package's Config struct hold the
ephemeralDevserver object if one is used. Also move
not-really-configuration fields to the end of the struct and
add a TODO to consider moving them somewhere else later.

I'm doing this to make it easier for a later change to
create a new ephemeralDevserver if the local test bundle
needs to be restarted after having been aborted by a bundle
or DUT crash.

BUG=chromium:889119,chromium:971517
TEST=manual: ran a test and saw that devserver was still
     initialized

Change-Id: Ic8835e0ea7a1093f0b11b511e524e3123facbd11