Update patch set 1
Patch Set 1: Code-Review-2

nya@, i'm uploading this more for conversation than as something that i'm sure i actually want to commit.

for https://crrev.com/c/1607126, it looks like we should create a new ephemeralDevserver when we restart testing (but maybe just when the SSH connection to the DUT is lost). right now, we create the ephemeralDevserver in local() and defer a call to close it when that function returns.

as seen here, i've considered moving ownership of the ephemeralDevserver into Config. that makes it possible to close and recreate it mid-testing. the ownership of the object feels confusing in this change, though, as it's created in local.go but closed in Config.Close. i'm also not sure that we want to let the ephemeralDevserver have the same lifetime as Config, as the DUT may be rebooted by remote tests.

i also considered making startEphemeralDevserver be a method on Config. this makes ownership clearer, but i'm not sure that Config should contain this sort of logic.

i might upload another version of this that introduces a new devserverConfig type, as that feels like it may be cleaner.

Patch-set: 1
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1115919 <1115919@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1000042 <1000042@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2