tast: Return a channel from doAsync.
Change doAsync function to return a channel to which the result
will be sent if ctx's deadline is reached before the function
finishes.

By returning a channel, I expect callers of doAsync to think
carefully about the timeout case, which is often important to
handle when the inner function calls constructors and has to
release allocated resources.

BUG=chromium:971024
TEST=fast_build.sh -T

Change-Id: I5fe91ce18832c4c10055743ff4e4e35e5e5432e7