timing: Support concurrent stages.
This change updates the timing package to be concurrency-safe.

Formerly, new stages were always added to the "active" stage
(the most recent not-yet-ended stages). This does not work for
concurrent use cases because two or more stages can be active
at a moment.

This change introduces the concept of the "current" stage.
A current stage is associated with context.Context, and
timing.Start adds a new stage to the current stage as a child.

BUG=chromium:970126
TEST=fast_build.sh -T
TEST=fast_build.sh && ~/go/bin/tast run betty arc.Boot
TEST=cat /tmp/tast/results/latest/timing.json

Cq-Depend: chromium:1651587
Change-Id: I4c4dcfd01162640dff7557d931c3b5797af1c22e
Reviewed-on: https://chromium-review.googlesource.com/1652873
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>