tast: Remove -checktestdeps=auto flag value.
-checktestdeps controls the behavior to skip tests by checking software
dependencies. Formerly the flag had three values: always, never, auto.
The auto value, which was actually the default if the flag is not
explicitly specified, indicates that software deps should be checked
if tests are specified by attribute expressions (e.g. "(!disabled)")
rather than exact names (e.g. "arc.Boot") or globs (e.g. "arc.*").

Since this behavior is confusing, this CL changes the flag to accept
a boolean value, and sets the default to always check software
dependencies.

Fortunately, there is no use of the flag in production today:
http://cs/search/?q=%22-checktestdeps%22

BUG=chromium:971486
TEST=fast_build.sh -T
TEST=fast_build.sh && ~/go/bin/tast run betty arc.Boot audio.DevicePlay

Change-Id: I3edb2eaca0ccd84f5826eb75362000d482906c27