tast: Support building multiple targets in parallel.
build.Build is updated to take []*build.Target each of which
describes a Go executable package to build, and it builds the
specified packages in parallel.

This change just introduces a new API in the build package.
We start to utilize the parallel build in the later CLs.

BUG=chromium:940761
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run -buildlocalrunner DUT example.Pass

Cq-Depend: chromium:1679798
Change-Id: I72bca27a9fa42dc5d3219f8aad4feb19fa8863cb