tast: Improve SSH's API to push files.
SSH.PutTreeRename used to take the base directory in the local and
remote host and files were specified in paths relative to the base
directory. Assumption here was that we usually want to push files
under a base directory, but it is not true today especially when
a test bundle binary and local_test_runner are pushed at once.

This CL changes the method to take just one map whose key is an
absolute/relative path to a local file and value is an absolute
path to a remote file. Since we no longer have concept of the base
directory, the method is renamed from PutTreeRename to PutFiles.

BUG=chromium:970127
BUG=chromium:970126
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=ssh DUT 'rm -rfv /usr/local/share/tast/data_pushed'
TEST=~/go/bin/tast run DUT example.DataFiles

Change-Id: Ic7d8783ccadaed6f002cc552911712c54671aa7f