tast: allow multiple tests per file
DO NOT SUBMIT

Demo of calling AddTest() multiple times in one file.

A test in "test_file.go" test must now *start* with "TestFile" but may be
longer. These additional tests will have a dot inserted in the test name.
For example, TestFileOne and TestFileTwo become "TestFile.One" and
"TestFile.Two".

The camel-case rule helps reject some common typos such since "TestFiles"
will be rejected (instead of becoming "TestFile.s"). We do not enforce
that multiple tests exist in a file before allowing "TestFile.Suffix" and
we allow a "TestFile" to exist even if there is also a "TestFile.Suffx".
This avoids forcing a a test to change name when other tests are added or
removed.

TEST=./fast_build.sh -T

Change-Id: Ia556a844d65ab58e8c7bee7bba0f47b616559e82