tast: Decide the output directory of local tests in the tast command.
Previously the output directory which local test bundles write
test artifacts to was decided by the local_test_runner and the
directory path was notified to the tast command by a RunEnd
control message sent at the end of all tests.

This is problematic in two points:
- When local_test_runner is aborted for timeout or crashes,
  RunEnd is not sent and thus it is impossible to locate and
  pull test artifacts.
- When local_test_runner is restarted for crashes, test
  artifacts are written to different directories every time.

This change makes the tast command always set OutDir in its
RunTests request (runner.Args.RunTests.BundleArgs.OutDir) by
generating a random directory name. Runners are also updated
to create the output directory if it does not exist.

For protocol compatibility, RunEnd.OutDir is still filled, and
the tast command issues a mkdir command via SSH to create the
output directory before running local_test_runner so that the
new tast command works well with older local_test_runner.

The change to fix the issue that test artifacts are not saved
when the global timeout is hit is coming next.

BUG=chromium:972457
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run -build=false -timeout=3 DUT arc.Boot
TEST=~/go/bin/tast run -build=true -timeout=3 DUT arc.Boot

Change-Id: I6f87676149722d5fc5fa2453b8b0c11109ef746e