tast: Pull output files even when runner aborts.
Previously output files from a runner were pulled on receiving a
RunEnd control message. This means output files were not pulled if
a runner crashed or timed out.

With this change, the tast command pulls output files when a runner
exits. OutDir in RunEnd is ignored by the tast command, but the
runner fills the field for a while for protocol compatibility.

BUG=chromium:972457
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run -timeout=15 DUT example.Perf ui.ChromeLogin

Change-Id: Ie7b9549951ca25dc765f9296e8f412c5a8df202a