tast: Do not clobber the temporary directory.
The directory passed to prepareTempDir is a base directory for
temporary files, just like /tmp. Thus it is incorrect to clobber
the directory since other processes might be using it concurrently.

When the temporary directory is not explicitly given in requests,
create a unique directory to avoid conflicts.

BUG=chromium:1000030
TEST=fast_build.sh -T
TEST=fast_build.sh
TEST=~/go/bin/tast run DUT example.Pass
TEST=ssh DUT rm -rf /usr/local/tmp/tast
TEST=~/go/bin/tast run DUT example.Pass

Change-Id: Ibf69e99783acb17490f3025fafdc4ee3d700aef9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/1782367
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>