tast: Fix a race condition in TestGetFileTimeout.
This test makes sure that Host.GetFile fails if the given context
is canceled during its execution.

Previously cancellation was performed by setting td.execTimeout
to endTimeout, which instructs testData.handleExec to cancel the
context just before a SSH command finishes. In the case of this
test, the SSH command is /bin/tar creating a tarball. However
actually its return code is ignored in GetFile; instead, exit code
of the local /bin/tar command connected via pipe is checked.

So, as opposed to the test's original intention, when the test
passed, the error was reported from the local /bin/tar due to
context cancellation. This is actually a race condition; the local
tar command can exit successfully before the context is canceled.
This race condition can be reproduced reliably by inserting
time.Sleep(time.Second) before td.cancel() in testData.handleExec.

This change fixes the race condition by using startTimeout which
cancels the context before a remote SSH command is executed. This
should preserve the original intention of the test.

BUG=chromium:1002191
TEST=fast_build.sh -T

Change-Id: I5430e2e79d11fb032a1695a6eac90497c0cb39f5