tast: Immediately unblock wait operations on Abort.
Previously Abort just called Session.Signal(SIGKILL) and hoped the
server to finish the command execution soon. However it is not
always the case and Session.Wait can block until the context
deadline. This is usually an undesired behavior.

With this change, Wait/Run/Output/CombinedOutput returns
immediately after calling Abort. SIGKILL is still sent to the
remote process (in waitAndClose), but the timing is changed to
when Wait/Run/Output/CombinedOutput is called.

BUG=chromium:971024
BUG=chromium:969627
TEST=fast_build.sh -T

Change-Id: I6dee0c9037a04b068063e9c0f51e7dd46749e849
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/1802717
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>