Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)
> 
> > Patch Set 4:
> > 
> > I don't think we should disable the entire test on sanitizer builds -- we do want to find memory errors in Minijail.
> > 
> > We should only disable the static part of the test in ASan builders.
> 
> Could you please let us know if there is a way to query if the current build is with asan in minijail.go? It seems the file runs both the dynamic and static version of each test, and the static tests seems all rely on security.Minijail.staticbashexec, which is statically linked. Do you mean to disable that test?

Doesn't the test know when any of the *San flags were passed? In that case, you would skip the static version of the test.

We should still run the dynamic version of all the tests when the test is running on an ASan builder.

Another option would be to skip the static tests if the test does not find the staticbashexec executable.

Patch-set: 5