tast: Fix a race around reading stderr of runners.
** Problem

When runLocalRunner sees a runner execution failure, it attempts
to read the first line of stderr and include it in the returned
error for better diagnosis. TestLocalExecFailure is the test to
ensure it is working. However we observed that sometimes it can
not read the first line of stderr:

local_test.go:284: local() logged "... Failed to run tests: Process
exited with status 1\n"; want substring "some failure message\n"

This failure was caused in the following race scenario within
runLocalRunner:

- local_test_runner is started with startLocalRunner
- firstLineReader is constructed with newFirstLineReader, but its
  background goroutine does not start immediately
- local_test_runner exits abnormally
- host.Cmd.Wait finishes with an error; at this point, handle.stderr
  is closed due to the contract of host.Cmd.Wait, which is almost
  similar to that of os/exec.Cmd.Wait:
  https://godoc.org/os/exec#Cmd.Wait
- The background goroutine of firstLineReader starts. It can't
  read anything from handle.stderr because it's already closed

Actually, as noted in the documentation of os/exec.Cmd.StdoutPipe,
it is incorrect to call Wait before finishing reading from pipes
created by StdoutPipe/StderrPipe:

> Wait will close the pipe after seeing the command exit, so most
> callers need not close the pipe themselves; however, an implication
> is that it is incorrect to call Wait before all reads from the pipe
> have completed.
https://godoc.org/os/exec#Cmd.StdoutPipe

** Solution

Stop using StderrPipe. Instead, set a buffer-like object to Cmd.Stderr.
Then we can read it after Wait reliably.

BUG=chromium:1003952
TEST=fast_build.sh -T
TEST=tools/go.sh test -count=1000 -run TestLocalExecFailure chromiumos/cmd/tast/run

Change-Id: Ieae7b38c5fe74877dcb863cbaa5004d415671ec3