Changing virtual_usb_printer to be gated on kernel version
The usbip kernel module is being changed to be installed on all
supported kernel versions using their kernel configs. This change
removes the use of the usbip use flag and instead checks to see that the
kernel version in use is not one of the unsupported versions.

BUG=chromium:943331
TEST=Passes virtual usb printer tast tests on atlas device

Cq-Depend: chromium:1924841, chromium:1924842, chromium:1925128
Cq-Depend: chromium:1924844
Change-Id: I28d8a630d0710e683c602c30ef6cfb32e54e3dc7