Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Non-lab case is a very good point. Yes, seems like -var is more appropriate.
> 
> I'm not sure it's an XOR (one or the other) case -- it's nice if Tast would do the hostname translation for us too, so we don't have to do 'tast run -var ...' every time. Even in non-lab cases, people don't always specify each router_addr, pcap_addr, etc., manually -- they rely on dnsname_mangler.py much of the time.

Thanks for the pointer to dnsname_mangler.py. I agree, it's convenient if we can run WiFi tests without explicit -var in typical cases.

Another approach would be to add a method to dut.DUT to return an SSH client for the companion device, i.e.

 func (d *dut.DUT) DefaultWifiHost() *dut.DUT

which constructs the wifi hostname from DUT hostname. Then tests can do:

 var wifi *dut.DUT
 if host, ok := s.Var("wifi.hostname"); ok {
   wifi = dut.New(host, d.KeyFile(), d.KeyDir())
 } else {
   wifi = d.DefaultWifiHost()
 }


> > Then in lab case we can pass extra vars in the Autotest control file for wifi tests.
> 
> For my curiosity: do we expect to have the Autotest control file as a harness here indefinitely? Or should Tast be aiming to be self-sufficient on this kind of stuff?

We'll move away from control files eventually, and it's possible that WiFi tests need to be modified to run on new infra in the future.

What I'm afraid is that random tests (not WiFi tests) may start to use real hostnames if we provide a way to get them. As long as we constraint the usage to WiFi tests for connecting to accompanied router only, the cost needed to migrate to the new infra should be limited. But if random tests depend on real hostnames I can't estimate the cost.

Patch-set: 3
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0