tast: Add an error message for debuggability.
When 'file' command is not available due to stateful partition not
being in the right state, the error message is not intuitive.

"2019/12/22 19:49:38 Failed to build and push: failed to get arch for
127.0.0.1:2224: Process exited with status 127"

If the output "file: command not found" was given I would have known
what happened.

BUG=None
TEST=ssh crbook sudo mv /usr/local/bin/file{,-}; cros_sdk tast run crbook arc.Boot
  # with output:
  # 2019/12/23 19:07:56 Failed to build and push: failed to get arch for
  #  localhost:2223: file command failed: Process exited with status 127
  #  (output: "bash: line 0: exec: file: not found\n")

Change-Id: I473c54b9f0fac706a22a8ee42b34c6bc9766fa61
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/1980103
Tested-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>