host: Use default user/port if not specified in NewSSH
As defaultSSHUser and defaultSSHPort are not exported, it'll be
convenient to have the default applied so that the callers don't need
to keep their copies of defaultSSHPort=22 or call ParseSSHTarget for
filling the default values.

The behavior will be used by companion devices in later commit.

BUG=chromium:1034875
TEST=fast_build.sh -T

Change-Id: I427fb4384544803d66a9f14ba913c403e6894113
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/1980108
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>