tast: Rename TestCase to TestInstance.
We used to have two very resembled types, Test and TestCase. Since
"case" does not add any meaningful explanation, it is essentially
impossible to tell the difference of the two types without actually
reading the code.

This change just renames TestCase to TestInstance to give some
meaningful explanation to the type.

I'm going to add more comments to describe the difference of two
types in a separate change.

This is a pure renaming change; no functional change is expected.

BUG=chromium:1036259
TEST=fast_build.sh -T
TEST=git grep -i testcase

Cq-Depend: chromium:1982000
Change-Id: I57060f3a5c892d28a930012ff2c3352f8691c9e4