Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3:
> 
> (2 comments)
> 
> Could you updat the doc, too?
> https://chromium.googlesource.com/chromiumos/platform/tast/+/HEAD/docs/test_dependencies.md
> 
> Also, the concept "crostini_cq" is somehow conflicting with the intention of software dependency,
> because it is designed to filter tests due to missing features, i.e., they wouldn't work by design.
> 
> R+=nya@ to hear more thoughts from him.

This seems to be the only way that Tast supports making this kind of distinction.  What I really want is the ability to make the test critical on some boards and informational on others, and to change the set of boards in each as tests stabilize or become flaky.

I actually started a design proposal along these lines but am still working on cleaning it up.

Please let me know if there is some other way to accomplish this, it is critical to the Crostini team to get some CQ test coverage, as things keep breaking without it.

I will update the doc if you are otherwise ok with the change.

Patch-set: 3
Reviewer: Gerrit User 1370239 <1370239@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>