errors: Add support of go1.13 errors.Unwrap
As go in chroot uprev to 1.13, let the wrapped errors in Tast support
go1.13 Unwrap interface. This is useful when we want to identify low
level cause of errors while keeping the wrapping messages on the call
path. In this case, we can call built-in errors.As or errors.Is instead
of matching Error() string.

This function is intended to be used for checking dbus errors that may
be wrapped.

BUG=chromium:1047568
TEST=fast_build.sh -T

Change-Id: I2278387773e7a92a23419ee8aa8ef54812b90240