errors: Add support of go1.13 errors.Unwrap
As go in chroot uprev to 1.13, let the wrapped errors in Tast support
go1.13 Unwrap interface. This is useful when we want to identify low
level cause of errors while keeping the wrapping messages on the call
path. In this case, we can call built-in errors.As or errors.Is instead
of matching Error() string.

This function is intended to be used for checking dbus errors that may
be wrapped.

BUG=chromium:1047568
TEST=fast_build.sh -T

Change-Id: I2278387773e7a92a23419ee8aa8ef54812b90240
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2032764
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>