Update patch set 2
Patch Set 2:

(2 comments)

> In thinking about this problem, I don't like that this CL breaks one of the rules of contexts - "Do not store Contexts inside a struct type; instead, pass a Context explicitly to each function that needs it."

We should pass context.Context as a function argument because context.Context is an immutable object that can be derived (with context.WithTimeout etc.) as we go through function calls.

In this case, ultimately, we have to pass two different contexts (prepare-scoped and precondition-scoped) to Prepare. But it's odd to pass two contexts at once as function arguments, e.g.

 func (p *Precondition) Prepare(ctx, longCtx context.Context, s *testing.State)

So we include the second context.Context in testing.State. We have an existing strong recommendation to avoid passing testing.State as a function argument, so we can assume that the longer context is not passed to another function via testing.State.


> But then again, Tast in general is already using contexts in an unintended way, since the contexts being passed around are actually TestContexts that overuses WithValue.  I'm not sure mixing two types of contexts is a good idea.
> 
> I'll add contexts as a topic for Monday's tast-reviewers sync.  Please hold off on submitting this for now.

That's a good point, I shared a note about use of context.Context in Tast to clarify.

Patch-set: 2
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>