Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > Thanks for your contribution. However I'm hesitant to add HTML reporting to Tast itself. It seems to me that it should be maintained outside of Tast.
> > 
> > Thanks for the review!
> > Usually, user will expect the report generation by framework itself after execution completed. If we keep the report generation outside the framework then it will be having one more dependency.
> 
> Any further comment on this :) or Can I go ahead and abandon the patch?

Sorry for delay, I missed your last comment on January.

I still think richer reporting is good to be done outside of Tast framework. Tast already outputs results in programatically parsable way (JSON) so it should be possible (if there's something missing we can add more data to JSON). And a good point about separating richer reporter is that we can apply the reporter to past results.

That said, I'm not 100% against adding HTML output if it is useful. But if we were to do that, I want the HTML generation code to be separated from text reporting code so that it's easier to extend.

Patch-set: 2
Reviewer: Gerrit User 1118346 <1118346@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>