Introduce video_overlays software dependency
Videos can be promoted to overlays on ChromeOS systems if the
appropriate pixel format is supported by the DRM planes. On amd64
systems, these coincide with those where we have drm_atomic. This
CL adds this software dependency so the video.MemCheck can use it
in crrev.com/c/2079653.

BUG=chromium:1047514
TEST=tast run ${IP} video.MemCheck.* on samus and atlas
samus skips and atlas runs correctly and as expected

Change-Id: I286ce4a36a0e0dab61238ea1544f5745c78dca23