tast: Protect the control message channel.
Test bundles communicate with the tast command on the host with the
control message channel established on stdin/stdout. Consequently,
if a test writes something to stdout or reads from stdin, the
control message channel is disturbed.

This change introduces bundle.{LocalDefault,RemoteDefault} which
closes os.Stdin, os.Stdout and os.Stderr before running tests.

BUG=chromium:1041131
TEST=fast_build.sh -T
TEST=tast run DUT meta.Stdio

Change-Id: Idfb7af3f1ac42a6cf7a82a1f409a6afe3f1c6c8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/2089160
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Keigo Oka <oka@chromium.org>