tast: Add "-skipsort" flag to tast command and bundle runner.
Running tests without sorting is important for certain performance
test when test cases must be run in the pre-defined order. And usually
test preconditions for these test cases are also pre-arranged
following the given test order and grouping tests with preconditions
is not desired either.

"tast run" is used to trigger a list of tests. Currently, the given
tests will be re-ordered (sorted with test names) before they are
actually run. This CL adds a "-skipsort" flag that can disable
the tests sorting when set to false.

To achieve the goal, this CL adds a new buildPlanSkipSort() function
to build a test plan ignoring the test Fixtures and let all tests
run sequentially with given order.

The "false" default value of the "-skipsort" flag makes sure when it
is not set the existing test behavior is not altered.

BUG=None
TEST=./fast_build.sh -T
TEST=tast run -skipsort=true <DUT> test1 test2 test3...

Change-Id: Icca16a07c5845b7550eef4e344edbfc4861da96f