tast: add more functions in state to support new precondition use cases.
- get precondition and reset precondition value
Once a precondition is set, there is a need to reset it's value. Eg: LoginReuse precondion
allows user to re-login and the precondition value (*chrome.Chrome) will need to be reset.
- get vars without registration
In order to use test variables in precondition, it's not necessary to register the
variables in each test cases with same precondition.

BUG=None
TEST=./fast_build.sh -T

Change-Id: Ieed1c9c82de8a92509e5d261957d946406e3d852