Add new dependency memfd_create
The memfd_create function was implemented in the 3.17 Linux kernel.
Exclude tests that rely on that function from boards with kernels
before 3.17.

Also, the (old) version of QEMU we are using doesn't have memfd_create
support, so exclude tests that rely on memfd_create from VMs.
(https://bugs.launchpad.net/qemu/+bug/1734792 is fixed but we're using a
version from before it was fixed.)

BUG=chromium:1060375
TEST=tast ... ui.ChromeCrashLoop*. Skipped on betty, runs on
veyron_minnie

Cq-Depend: chromium:2124953
Change-Id: Ic9f777f445df082ed1d4ae72ab7b82c62c39a948