Tast: Add a sub-attribute "unstable" under "group:wificell".
A test with attribute {"group:wificell", wifi_func, unstable} means it
is a test verifying WiFi basic functionality which should run at a
wificell fixture and it is not stable.

A test with attribute {"group:wificell", wifi_func} means it is a test
like above except it is stable. That means it is part of the suite
suite:wifi_matfunc.

BUG=chromium:1034878
TEST=fast_build.sh -T

Change-Id: I89e9dedd0c4e9c20a830beca6b53c8e8a5c80faf